[osquery] flatten and support last in non_office_logins
