+++ osquery_rules/osquery_linux_aws_commands.py
