+++ osquery_rules/osquery_mac_application_firewall.py
