+++ osquery_rules/osquery_mac_enable_auto_update.py
