+++ osquery_rules/osquery_mac_osx_attacks.yml
