+++ osquery_rules/osquery_mac_osx_attacks_keyboard_events.yml
