+++ osquery_rules/osquery_mac_unwanted_chrome_extensions.py
