+++ osquery_rules/osquery_ossec.py
