+++ osquery_rules/osquery_ossec.yml
