+++ osquery_rules/osquery_suspicious_cron.py
